+++ src/filters/BaseFilter.ts
